Changed Native Quadtree coordinates from float to int, for a modest 10-20% speedup

+++ SDNative/quadtree/QuadTreeCAPI.cpp
@@ -6,7 +6,7 @@ using tree::QtreeNode;
 using tree::QtreeBoundedNode;
 using tree::SpatialObj;
 
-DLLEXPORT QuadTree* __stdcall QtreeCreate(float universeSize, float smallestCell)
+DLLEXPORT QuadTree* __stdcall QtreeCreate(int universeSize, int smallestCell)
 {
     return new QuadTree(universeSize, smallestCell);
 }
